backend/pyproject.toml
[project]
name = "ai-assistant-showcase"
version = "0.1.0"
description = "AI assistant for product and FAQ consultations"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "sqlalchemy[asyncio]",
    "alembic",
    "asyncpg[binary]",
    "pydantic-settings",
    "python-multipart",
    "qdrant-client>=1.17.1",
    "sentence-transformers>=5.4.1",
    "torch>=2.11.0",
    "litellm>=1.79.1",
    "aiogram>=3.22.0",
]

[project.optional-dependencies]
dev = ["pytest", "pytest-cov", "pytest-asyncio", "httpx", "ruff", "faker"]

[tool.uv.sources]
torch = { index = "pytorch-cpu" }

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
addopts = "--cov=. --cov-report=term-missing --cov-fail-under=80"
asyncio_mode = "auto"

[tool.coverage.run]
omit = ["*/__init__.py", "tests/*"]

[tool.coverage.report]
exclude_lines = ["pragma: no cover", "if TYPE_CHECKING:"]

[tool.ruff]
line-length = 110

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.ruff.lint.per-file-ignores]
"alembic/*" = ["E402"]
"tests/conftest.py" = ["F403"]

[tool.alembic]
script_location = "%(here)s/alembic"
prepend_sys_path = [
    "."
]
timezone = "UTC"
output_encoding = "utf-8"

# [[tool.alembic.post_write_hooks]]
# lint with attempts to fix using "ruff" - use the module runner, against the "ruff" module
# name = "ruff"
# type = "module"
# module = "ruff"
# options = "check --fix REVISION_SCRIPT_FILENAME"
